refactor: split dashboard page into components
